Add request id in middleware
A request now has a unique tyk-content-request-id which is truely unique
to every request.  There is also a request-id which you can pass
through.  If blank the request-id will default to the
tyk-content-request-id (as a generation method on the initial request).

Have had to add tyk.js, which is apparently not needed so should remove
this in the future. (see
https://github.com/TykTechnologies/tyk/commit/0f5ab6ce9500efb6845749f0abe5acdfaa25f2fd)